balena CI: Add balena-cli executable signing step
Change-type: minor
