+++ package.json
@@ -72,7 +72,8 @@
     "bin": "balena",
     "commands": "./build/actions-oclif",
     "macos": {
-      "identifier": "io.balena.cli"
+      "identifier": "io.balena.cli",
+      "sign": "Developer ID Installer: Rulemotion Ltd (66H43P8FRG)"
     }
   },
   "devDependencies": {
